IDS_GraphQL/requirements.txt
Django
graphene
graphene-django
graphql-core
graphql-relay
psycopg2
dj-database-url
graphene-file-upload
pillow
google-generativeai 
django-graphql-jwt
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client 
